@@ -1,6 +1,4 @@
-/// <reference path="LeapMotionTS.d.ts" />
-
-import Leap = require('LeapMotionTS');
+import Leap = require('./LeapMotionTS');
 
 var controller: Leap.Controller = new Leap.Controller();
 controller.addEventListener(Leap.LeapEvent.LEAPMOTION_FRAME, (event: Leap.LeapEvent) => {
