@@ -30,11 +30,11 @@ declare module angular.animate {
         /**
         * Globally enables / disables animations.
         *
-        * @param element If provided then the element will be used to represent the enable/disable operation.
         * @param value If provided then set the animation on or off.
+        * @param element If provided then the element will be used to represent the enable/disable operation.
         * @returns current animation state
         */
-        enabled(element?: JQuery, value?: boolean): boolean;
+        enabled(value?: boolean, element?: JQuery): boolean;
 
         /**
          * Performs an inline animation on the element.
