@@ -1,4 +1,4 @@
-/// <reference path="bluebird.d.ts" />
+///<reference path="bluebird.d.ts"/>
 
 // Tests by: Bart van der Schoor <https://github.com/Bartvds>
 
@@ -369,12 +369,12 @@ fooProm = fooProm.timeout(num, str);
 // - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 
 fooProm.nodeify();
-fooProm = fooProm.nodeify((err: any) => {
+fooProm = fooProm.nodeify((err: any) => { });
+fooProm = fooProm.nodeify((err: any, foo?: Foo) => { });
 
-});
-fooProm = fooProm.nodeify((err: any, foo?: Foo) => {
-
-});
+fooProm.nodeify({ spread: true });
+fooProm = fooProm.nodeify((err: any) => { }, { spread: true });
+fooProm = fooProm.nodeify((err: any, foo?: Foo) => { }, { spread: true });
 
 // - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 
@@ -508,6 +508,17 @@ barArrProm = fooProm.map<Foo, Bar>((item: Foo) => {
 	return bar;
 });
 
+barArrProm = fooProm.map<Foo, Bar>((item: Foo, index: number, arrayLength: number) => {
+	return bar;
+}, {
+	concurrency: 1
+});
+barArrProm = fooProm.map<Foo, Bar>((item: Foo) => {
+	return bar;
+}, {
+	concurrency: 1
+});
+
 // - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 
 barProm = fooProm.reduce<Foo, Bar>((memo: Bar, item: Foo, index: number, arrayLength: number) => {
@@ -524,6 +535,17 @@ fooArrProm = fooArrProm.filter<Foo>((item: Foo, index: number, arrayLength: numb
 });
 fooArrProm = fooArrProm.filter<Foo>((item: Foo) => {
 	return bool;
+});
+
+fooArrProm = fooArrProm.filter<Foo>((item: Foo, index: number, arrayLength: number) => {
+	return bool;
+}, {
+	concurrency: 1
+});
+fooArrProm = fooArrProm.filter<Foo>((item: Foo) => {
+	return bool;
+}, {
+	concurrency: 1
 });
 
 // - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
@@ -617,10 +639,39 @@ voidProm = Promise.delay(num);
 
 func = Promise.promisify(f);
 func = Promise.promisify(f, obj);
-;
 
 obj = Promise.promisifyAll(obj);
 anyProm = Promise.fromNode(callback => nodeCallbackFunc(callback));
+
+// - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+
+declare var util: any;
+
+function defaultFilter(name: string, func: Function) {
+    return util.isIdentifier(name) &&
+        name.charAt(0) !== "_" &&
+        !util.isClass(func);
+}
+
+function DOMPromisifier(originalMethod: Function) {
+    // return a function
+    return function promisified() {
+        var args = [].slice.call(arguments);
+        // Needed so that the original method can be called with the correct receiver
+        var self = this;
+        // which returns a promise
+        return new Promise(function(resolve, reject) {
+            args.push(resolve, reject);
+            originalMethod.apply(self, args);
+        });
+    };
+}
+
+obj = Promise.promisifyAll(obj, {
+	suffix: "",
+	filter: defaultFilter,
+	promisifier: DOMPromisifier
+});
 
 // - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 
@@ -709,6 +760,26 @@ barArrProm = Promise.map(fooThenArrThen, (item: Foo, index: number, arrayLength:
 	return barThen;
 });
 
+barArrProm = Promise.map(fooThenArrThen, (item: Foo) => {
+	return bar;
+}, {
+	concurrency: 1
+});
+barArrProm = Promise.map(fooThenArrThen, (item: Foo) => {
+	return barThen;
+}, {
+	concurrency: 1
+});
+barArrProm = Promise.map(fooThenArrThen, (item: Foo, index: number, arrayLength: number) => {
+	return bar;
+}, {
+	concurrency: 1
+});
+barArrProm = Promise.map(fooThenArrThen, (item: Foo, index: number, arrayLength: number) => {
+	return barThen;
+}, {
+	concurrency: 1
+});
 // - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 
 // fooArrThen
@@ -724,6 +795,27 @@ barArrProm = Promise.map(fooArrThen, (item: Foo, index: number, arrayLength: num
 });
 barArrProm = Promise.map(fooArrThen, (item: Foo, index: number, arrayLength: number) => {
 	return barThen;
+});
+
+barArrProm = Promise.map(fooArrThen, (item: Foo) => {
+	return bar;
+}, {
+	concurrency: 1
+});
+barArrProm = Promise.map(fooArrThen, (item: Foo) => {
+	return barThen;
+}, {
+	concurrency: 1
+});
+barArrProm = Promise.map(fooArrThen, (item: Foo, index: number, arrayLength: number) => {
+	return bar;
+}, {
+	concurrency: 1
+});
+barArrProm = Promise.map(fooArrThen, (item: Foo, index: number, arrayLength: number) => {
+	return barThen;
+}, {
+	concurrency: 1
 });
 
 // - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
@@ -743,6 +835,27 @@ barArrProm = Promise.map(fooThenArr, (item: Foo, index: number, arrayLength: num
 	return barThen;
 });
 
+barArrProm = Promise.map(fooThenArr, (item: Foo) => {
+	return bar;
+}, {
+	concurrency: 1
+});
+barArrProm = Promise.map(fooThenArr, (item: Foo) => {
+	return barThen;
+}, {
+	concurrency: 1
+});
+barArrProm = Promise.map(fooThenArr, (item: Foo, index: number, arrayLength: number) => {
+	return bar;
+}, {
+	concurrency: 1
+});
+barArrProm = Promise.map(fooThenArr, (item: Foo, index: number, arrayLength: number) => {
+	return barThen;
+}, {
+	concurrency: 1
+});
+
 // - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 
 // fooArr
@@ -758,6 +871,27 @@ barArrProm = Promise.map(fooArr, (item: Foo, index: number, arrayLength: number)
 });
 barArrProm = Promise.map(fooArr, (item: Foo, index: number, arrayLength: number) => {
 	return barThen;
+});
+
+barArrProm = Promise.map(fooArr, (item: Foo) => {
+	return bar;
+}, {
+	concurrency: 1
+});
+barArrProm = Promise.map(fooArr, (item: Foo) => {
+	return barThen;
+}, {
+	concurrency: 1
+});
+barArrProm = Promise.map(fooArr, (item: Foo, index: number, arrayLength: number) => {
+	return bar;
+}, {
+	concurrency: 1
+});
+barArrProm = Promise.map(fooArr, (item: Foo, index: number, arrayLength: number) => {
+	return barThen;
+}, {
+	concurrency: 1
 });
 
 // - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
@@ -853,6 +987,27 @@ fooArrProm = Promise.filter(fooThenArrThen, (item: Foo, index: number, arrayLeng
 	return boolThen;
 });
 
+fooArrProm = Promise.filter(fooThenArrThen, (item: Foo) => {
+	return bool;
+}, {
+	concurrency: 1
+});
+fooArrProm = Promise.filter(fooThenArrThen, (item: Foo) => {
+	return boolThen;
+}, {
+	concurrency: 1
+});
+fooArrProm = Promise.filter(fooThenArrThen, (item: Foo, index: number, arrayLength: number) => {
+	return bool;
+}, {
+	concurrency: 1
+});
+fooArrProm = Promise.filter(fooThenArrThen, (item: Foo, index: number, arrayLength: number) => {
+	return boolThen;
+}, {
+	concurrency: 1
+});
+
 // - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 
 // fooArrThen
@@ -868,6 +1023,27 @@ fooArrProm = Promise.filter(fooArrThen, (item: Foo, index: number, arrayLength: 
 });
 fooArrProm = Promise.filter(fooArrThen, (item: Foo, index: number, arrayLength: number) => {
 	return boolThen;
+});
+
+fooArrProm = Promise.filter(fooArrThen, (item: Foo) => {
+	return bool;
+}, {
+	concurrency: 1
+});
+fooArrProm = Promise.filter(fooArrThen, (item: Foo) => {
+	return boolThen;
+}, {
+	concurrency: 1
+});
+fooArrProm = Promise.filter(fooArrThen, (item: Foo, index: number, arrayLength: number) => {
+	return bool;
+}, {
+	concurrency: 1
+});
+fooArrProm = Promise.filter(fooArrThen, (item: Foo, index: number, arrayLength: number) => {
+	return boolThen;
+}, {
+	concurrency: 1
 });
 
 // - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
@@ -887,6 +1063,27 @@ fooArrProm = Promise.filter(fooThenArr, (item: Foo, index: number, arrayLength: 
 	return boolThen;
 });
 
+fooArrProm = Promise.filter(fooThenArr, (item: Foo) => {
+	return bool;
+}, {
+	concurrency: 1
+});
+fooArrProm = Promise.filter(fooThenArr, (item: Foo) => {
+	return boolThen;
+}, {
+	concurrency: 1
+});
+fooArrProm = Promise.filter(fooThenArr, (item: Foo, index: number, arrayLength: number) => {
+	return bool;
+}, {
+	concurrency: 1
+});
+fooArrProm = Promise.filter(fooThenArr, (item: Foo, index: number, arrayLength: number) => {
+	return boolThen;
+}, {
+	concurrency: 1
+});
+
 // - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 
 // fooArr
@@ -902,6 +1099,27 @@ fooArrProm = Promise.filter(fooArr, (item: Foo, index: number, arrayLength: numb
 });
 fooArrProm = Promise.filter(fooArr, (item: Foo, index: number, arrayLength: number) => {
 	return boolThen;
+});
+
+fooArrProm = Promise.filter(fooArr, (item: Foo) => {
+	return bool;
+}, {
+	concurrency: 1
+});
+fooArrProm = Promise.filter(fooArr, (item: Foo) => {
+	return boolThen;
+}, {
+	concurrency: 1
+});
+fooArrProm = Promise.filter(fooArr, (item: Foo, index: number, arrayLength: number) => {
+	return bool;
+}, {
+	concurrency: 1
+});
+fooArrProm = Promise.filter(fooArr, (item: Foo, index: number, arrayLength: number) => {
+	return boolThen;
+}, {
+	concurrency: 1
 });
 
 // - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
