@@ -1,3 +1,5 @@
+/// <reference path="./q-retry.d.ts" />
+
 ﻿import Q = require('q-retry');
 
 Q
@@ -9,7 +11,7 @@ Q
         return 0;
     })
     .retry(num => {
-        num.toFixed;    
+        num.toFixed;
     })
     .retry(() => {
 
@@ -32,7 +34,7 @@ Q
         limit: 10,
         interval: 1000,
         maxInterval: 20000,
-        intervalMultiplier: 1.5    
+        intervalMultiplier: 1.5
     })
     .then(str => {
         str.charAt;
