@@ -117,24 +117,24 @@ var promiseOrResponse = interceptor({
 });
 
 client = rest
-    .wrap(defaultRequest)
-    .wrap(hateoas)
-    .wrap(location)
-    .wrap(mime)
-    .wrap(pathPrefix)
-    .wrap(basicAuth)
-    .wrap(oAuth)
-    .wrap(csrf)
-    .wrap(errorCode)
-    .wrap(retry)
-    .wrap(timeout)
-    .wrap(jsonp)
-    .wrap(xdomain)
-    .wrap(xhr)
-    .wrap(noop)
-    .wrap(fail)
-    .wrap(knownConfig, { prop: 'value' })
-    .wrap(transformedConfig, { prop: 'value' });
+    .wrap<any>(defaultRequest)
+    .wrap<any>(hateoas)
+    .wrap<any>(location)
+    .wrap<any>(mime)
+    .wrap<any>(pathPrefix)
+    .wrap<any>(basicAuth)
+    .wrap<any>(oAuth)
+    .wrap<any>(csrf)
+    .wrap<any>(errorCode)
+    .wrap<any>(retry)
+    .wrap<any>(timeout)
+    .wrap<any>(jsonp)
+    .wrap<any>(xdomain)
+    .wrap<any>(xhr)
+    .wrap<any>(noop)
+    .wrap<any>(fail)
+    .wrap<any>(knownConfig, { prop: 'value' })
+    .wrap<any>(transformedConfig, { prop: 'value' });
 
 import xhrClient = require('rest/client/xhr');
 import nodeClient = require('rest/client/node');
