@@ -143,6 +143,20 @@ declare module createjs {
         {
 
         }
+
+    export class PlayPropsConfig
+	{
+	delay:number;
+	duration:number;
+	interrupt:string;
+	loop:number;
+	offset:number;
+	pan:number;
+	startTime:number;
+	volume:number;
+	static create( value:PlayPropsConfig|any ): PlayPropsConfig;
+	set ( props:any ): PlayPropsConfig;
+	}
     
     export class Sound extends EventDispatcher
         {
@@ -161,7 +175,9 @@ declare module createjs {
         static PLAY_INTERRUPTED: string;
         static PLAY_SUCCEEDED: string;
         static SUPPORTED_EXTENSIONS: string[];
-        
+	static muted: boolean;
+	static volume: number;
+        static capabilities: any;
         
         // methods
         static createInstance(src: string): AbstractSoundInstance;
