@@ -184,7 +184,7 @@
         geometry.computeBoundingSphere();
 
         var material = new THREE.MeshPhongMaterial({
-            color: 0xaaaaaa, ambient: 0xaaaaaa, specular: 0xffffff, shininess: 250,
+            color: 0xaaaaaa, specular: 0xffffff, shininess: 250,
             side: THREE.DoubleSide, vertexColors: THREE.VertexColors
         });
 
