@@ -48,7 +48,7 @@
         var s = 250;
 
         var cube = new THREE.BoxGeometry(s, s, s);
-        var material = new THREE.MeshPhongMaterial({ ambient: 0x333333, color: 0xffffff, specular: 0xffffff, shininess: 50 });
+        var material = new THREE.MeshPhongMaterial({ color: 0xffffff, specular: 0xffffff, shininess: 50 });
 
 
         for (var i = 0; i < 3000; i++) {
