@@ -80,7 +80,7 @@
         // GROUND
 
         var groundGeo = new THREE.PlaneBufferGeometry(10000, 10000);
-        var groundMat = new THREE.MeshPhongMaterial({ ambient: 0xffffff, color: 0xffffff, specular: 0x050505 });
+        var groundMat = new THREE.MeshPhongMaterial({ color: 0xffffff, specular: 0x050505 });
         groundMat.color.setHSL(0.095, 1, 0.75);
 
         var ground = new THREE.Mesh(groundGeo, groundMat);
