@@ -21,7 +21,7 @@ declare module "vinyl-fs" {
       cwd?: string;
 
       /**
-      * Specifies the folder relative to the cwd 
+      * Specifies the folder relative to the cwd
       * This is used to determine the file names when saving in .dest()
       * Default is where the glob begins
       */
@@ -34,16 +34,16 @@ declare module "vinyl-fs" {
       */
       buffer?: boolean;
 
-      /** 
+      /**
       * Setting this to false will ignore the contents of the file and disable writing to disk to speed up operations
-      * Defaults to true 
+      * Defaults to true
       */
       read?: boolean;
 
       /**  Only find files that have been modified since the time specified */
       since?: Date|number;
 
-      /** Setting this to true will create a duplex stream, one that passes through items and emits globbed files. 
+      /** Setting this to true will create a duplex stream, one that passes through items and emits globbed files.
       * Defaults to false */
       passthrough?: boolean;
 
@@ -63,7 +63,7 @@ declare module "vinyl-fs" {
 
    /**
    * This is just a glob-watcher
-   * 
+   *
    * @param globs Takes a glob string or an array of glob strings as the first argument
    * Globs are executed in order, so negations should follow positive globs
    * fs.src(['!b*.js', '*.js']) would not exclude any files, but this would: fs.src(['*.js', '!b*.js'])
@@ -72,7 +72,7 @@ declare module "vinyl-fs" {
 
    /**
    * This is just a glob-watcher
-   * 
+   *
    * @param globs Takes a glob string or an array of glob strings as the first argument
    * Globs are executed in order, so negations should follow positive globs
    * fs.src(['!b*.js', '*.js']) would not exclude any files, but this would: fs.src(['*.js', '!b*.js'])
@@ -89,8 +89,8 @@ declare module "vinyl-fs" {
    * @param folder destination folder
    */
    function dest(folder: string, opt?: {
-      /** Specify the working directory the folder is relative to 
-      * Default is process.cwd() 
+      /** Specify the working directory the folder is relative to
+      * Default is process.cwd()
       */
       cwd?: string;
 
@@ -125,14 +125,17 @@ declare module "vinyl-fs" {
    * cwd, base, and path will be overwritten to match the folder
    */
    function symlink(folder: string, opts?: {
-      /** 
-      * Specify the working directory the folder is relative to 
+      /**
+      * Specify the working directory the folder is relative to
       * Default is process.cwd()
       */
       cwd?: string;
 
-      /** 
-      * Specify the mode the directory should be created with 
+      /** Specify the mode the directory should be created with. Default is the process mode */
+      mode?: number|string;
+
+      /**
+      * Specify the mode the directory should be created with
       * Default is the process mode
       */
       dirMode?: number
@@ -146,14 +149,14 @@ declare module "vinyl-fs" {
    */
    function symlink(getFolderPath: (File: File) => string, opts?:
       {
-         /** 
-         * Specify the working directory the folder is relative to 
+         /**
+         * Specify the working directory the folder is relative to
          * Default is process.cwd()
          */
          cwd?: string;
 
-         /** 
-         * Specify the mode the directory should be created with 
+         /**
+         * Specify the mode the directory should be created with
          * Default is the process mode
          */
          dirMode?: number
